refactor: cleaned up main.rs
