+++ debug.log
@@ -1,1 +1,1 @@
-[perf avg over 60f] cpu_total=16.6ms acquire=0.0 record=0.1 submit=2.6 present=0.0 update=6.8 | gpu_total=6.3ms sky=3.6 phase1=1.2 pyramid=1.3 phase2=0.3 ui=0.0 | resident=2200
+[perf avg over 60f] cpu_total=16.6ms acquire=0.0 record=0.1 submit=2.6 present=0.0 update=6.2 | gpu_total=7.0ms sky=4.4 phase1=0.8 pyramid=1.5 phase2=0.3 ui=0.0 | resident=2185
