+++ src/main.rs
@@ -1,25 +1,23 @@
-#![allow(clippy::too_many_arguments)]
 mod game_state;
+mod event_handling;
+mod menu;
+mod hud;
+mod utils;
 mod graphical_core;
 mod storage;
 mod voxel;
 use anyhow::Result;
-use game_state::GameState;
-use graphical_core::{
-    camera::{Camera, EyeMatrices}, 
-    input::InputState, 
-    ui_pipeline::UiPipeline, 
-    vulkan_object::VulkanApplication
+use event_handling::{
+    initialize_event_info, 
+    handle_device_event, 
+    handle_window_event, 
+    handle_wait_event, 
+    release_cursor
 };
-use std::time::Instant;
-use voxel::player::Player;
 use vulkan_rust::{vk, Version};
 use winit::{
-    dpi::LogicalSize,
-    event::{DeviceEvent, ElementState, Event, MouseButton, WindowEvent},
-    event_loop::{EventLoop, EventLoopWindowTarget},
-    keyboard::{KeyCode, PhysicalKey},
-    window::{CursorGrabMode, Window, WindowBuilder},
+    event::Event,
+    event_loop::EventLoop
 };
 
 const PORTABILITY_MACOS_VERSION: Version = Version::new(1, 3, 216);
@@ -27,58 +25,6 @@ const VALIDATION_ENABLED: bool = cfg!(debug_assertions);
 
 const VALIDATION_LAYER: &std::ffi::CStr = c"VK_LAYER_KHRONOS_validation";
 const DEVICE_EXTENSIONS: &[&std::ffi::CStr] = &[vk::extension_names::KHR_SWAPCHAIN_EXTENSION_NAME];
-
-const BUTTON_W: f32 = 300.0;
-const BUTTON_H: f32 = 40.0;
-const TEXT_SIZE: f32 = 20.0;
-const TITLE_SIZE: f32 = 48.0;
-const BUTTON_COLOR: [f32; 4] = [0.2, 0.2, 0.3, 0.85];
-const BUTTON_HOVER: [f32; 4] = [0.3, 0.3, 0.5, 0.9];
-const TEXT_COLOR: [f32; 4] = [1.0, 1.0, 1.0, 1.0];
-const DIM_TEXT: [f32; 4] = [0.7, 0.7, 0.7, 1.0];
-
-struct EventInfo {
-    game_state: GameState,
-    input: InputState,
-    player: Player,
-    camera: Camera,
-    fps_counter: FpsCounter,
-    last_frame: Instant,
-    application: VulkanApplication,
-    user_window: Window,
-    cursor_position: [f32; 2],
-    physics_accumulator: f32,
-    minimized: bool,
-    menu_click: bool,
-    destroy_application: bool,
-}
-
-fn initialize_event_info(event_handler: &EventLoopWindowTarget<()>) -> anyhow::Result<EventInfo> {
-
-    let user_window = WindowBuilder::new()
-        .with_title("Manifold")
-        .with_inner_size(LogicalSize::new(1024, 768))
-        .build(&event_handler)?;
-
-    let application = unsafe { VulkanApplication::create_vulkan_application(&user_window) }?;
-    
-    let event_info= EventInfo {
-        game_state: GameState::TitleScreen,
-        input: InputState::new(),
-        player: Player::new(),
-        camera: Camera::default(),
-        fps_counter: FpsCounter::new(),
-        last_frame: Instant::now(),
-        application,
-        user_window,
-        cursor_position: [0.0; 2],
-        physics_accumulator: 0.0,
-        minimized: false,
-        menu_click: false,
-        destroy_application: false
-    };
-    Ok(event_info)
-}
 
 fn main() -> Result<()> {
         
@@ -99,462 +45,6 @@ fn main() -> Result<()> {
     Ok(())
 }
 
-fn handle_window_event(event: WindowEvent, event_info: &mut EventInfo, current_window: &EventLoopWindowTarget<()>) -> anyhow::Result<()> {
-    match event {
-        WindowEvent::CloseRequested => {
-            exit_program(&mut event_info.destroy_application, current_window, &mut event_info.application);
-        }
-        WindowEvent::Resized(size) => {
-            if size.width == 0 || size.height == 0 {
-                event_info.minimized = true;
-            } else {
-                event_info.minimized = false;
-                event_info.application.resized = true;
-            }
-        }
-        WindowEvent::CursorMoved { position, .. } => {
-            event_info.cursor_position = [position.x as f32, position.y as f32];
-        }
-        WindowEvent::KeyboardInput {..} => handle_keyboard_input(),
-        WindowEvent::MouseInput { state: ElementState::Pressed, button: MouseButton::Left, .. } => {
-            if event_info.game_state.is_menu() {
-                event_info.menu_click = true;
-            } else {
-                event_info.input.mouse_pressed(MouseButton::Left);
-            }
-        }
-        WindowEvent::MouseInput {state: ElementState::Pressed, button, .. } => {
-            if !event_info.game_state.is_menu() {
-                event_info.input.mouse_pressed(button);
-            }
-        }
-        WindowEvent::RedrawRequested {} => {
-            if event_info.destroy_application || event_info.minimized {
-                return Ok(());
-            }
-            let eyes = EyeMatrices::from_camera(&event_info.camera, event_info.application.swapchain_extent());
-            let result = match &event_info.game_state {
-                GameState::Playing => draw_ui(event_info, eyes),
-                _ => {
-                    let extent = event_info.application.swapchain_extent();
-                    let screen_width = extent.width as f32;
-                    let screen_height = extent.height as f32;
-                    let clicked = event_info.menu_click;
-                    event_info.menu_click = false;
-
-                    event_info.application.ui.begin_frame();
-                    draw_menu(&mut event_info.application.ui, &mut event_info.game_state, screen_width, screen_height, event_info.cursor_position, clicked);
-                    unsafe { event_info.application.render_menu_frame(&event_info.user_window, &eyes) }
-                }
-            };
-            if let Err(e) = result {
-                eprintln!("Render error: {e}");
-                exit_program(&mut event_info.destroy_application, current_window, &mut event_info.application);
-            }
-        }
-        _ => {}
-    }
-    Ok(())
-}
-
-
-fn draw_ui(event_info: &mut EventInfo, eyes: EyeMatrices) -> anyhow::Result<()> {
-    event_info.application.ui.begin_frame();
-    event_info.application.ui.draw_text(&event_info.fps_counter.display(), 4.0, 4.0, 16.0, [1.0, 1.0, 1.0, 0.8]);
-    let pos_text = format!("pos=({:.1},{:.1s},{:.1})", event_info.player.x, event_info.player.y, event_info.player.z);
-    event_info.application.ui.draw_text(&pos_text, 4.0, 22.0, 14.0, [1.0, 1.0, 1.0, 0.8]);
-    unsafe { event_info.application.render_frame(&event_info.user_window, &event_info.camera, &eyes) }
-}
-
-fn handle_device_event(event: DeviceEvent, event_info: &mut EventInfo) {
-    match event {
-        DeviceEvent::MouseMotion { delta: (dx, dy) } => {
-            if matches!(event_info.game_state, GameState::Playing) {
-                event_info.input.accumulate_mouse_delta(dx, dy);
-            }
-        }
-         _ => ()
-    }
-}
-
-fn handle_keyboard_input() {
-    todo!();
-}
-
-fn handle_wait_event(event_info: &mut EventInfo) {
-    const MAX_PHYSICS_CATCHUP: f32 = 0.25;
-    const PHYSICS_TICK: f32 = 1.0 / 60.0;
-    let now = Instant::now();
-    let delta_time = (now - event_info.last_frame).as_secs_f32();
-    event_info.last_frame = now;
-    event_info.fps_counter.tick(delta_time);
-    match &mut event_info.game_state {
-        GameState::Playing => {
-            event_info.input.apply_mouse_look(&mut event_info.player);
-            event_info.physics_accumulator = (event_info.physics_accumulator + delta_time).min(MAX_PHYSICS_CATCHUP);
-            while event_info.physics_accumulator >= PHYSICS_TICK {
-                    if let Some(world) = event_info.application.world() {
-                        let local_p = world.metric.sample_metric_at_pos(event_info.player.world_position()).minkowski_exponent;
-                        event_info.input.tick_movement(&mut event_info.player, world, PHYSICS_TICK, local_p);
-                    }
-                    event_info.physics_accumulator -= PHYSICS_TICK;
-                }
-                event_info.camera.sync_from_player(&event_info.player);
-                event_info.input.take_left_click();
-                event_info.input.take_right_click();
-            }
-            GameState::PreGenerating { .. } => {
-                unsafe { application.update_world(&camera).ok() };
-                tick_pregen(&mut game_state, &mut application, &user_window, &mut camera, &mut player);
-            }
-            GameState::EnteringWorld { .. } => {
-                if let GameState::EnteringWorld { world_dir, seed } = std::mem::replace(&mut game_state, GameState::Playing) {
-                    if let Err(e) = unsafe { application.enter_world(&world_dir, seed) } {
-                        eprintln!("Failed to enter world: {e}");
-                        game_state = GameState::TitleScreen;
-                    } else {
-                        grab_cursor(&user_window);
-                        camera = Camera::default();
-                        player = Player::new();
-                    }
-                }
-            }
-            _ => {}
-        }
-        user_window.request_redraw();
-}
-
-fn temp() {
-    event_handler
-        .run(move |event, current_window| match event {
-            Event::WindowEvent {
-                event: WindowEvent::KeyboardInput { event: key_event, .. },
-                ..
-            } => {
-                if let PhysicalKey::Code(key_code) = key_event.physical_key {
-                    match &mut game_state {
-                        GameState::Playing => match key_event.state {
-                            ElementState::Pressed => {
-                                if key_code == KeyCode::Escape {
-                                    unsafe { application.exit_world() };
-                                    game_state = GameState::TitleScreen;
-                                    release_cursor(&user_window);
-                                    return;
-                                }
-                                if key_code == KeyCode::KeyF && !key_event.repeat {
-                                    player.toggle_fly_mode();
-                                }
-                                if key_code == KeyCode::Space {
-                                    player.jump();
-                                }
-                                input.key_pressed(key_code);
-                            }
-                            ElementState::Released => input.key_released(key_code),
-                        },
-                        GameState::CreateWorld { name, seed_text } => {
-                            if key_event.state == ElementState::Pressed {
-                                if key_code == KeyCode::Escape {
-                                    game_state = GameState::TitleScreen;
-                                    return;
-                                }
-                                if key_code == KeyCode::Backspace {
-                                    if !seed_text.is_empty() {
-                                        seed_text.pop();
-                                    } else {
-                                        name.pop();
-                                    }
-                                    return;
-                                }
-                                if key_code == KeyCode::Enter && !name.is_empty() {
-                                    let seed: u32 = seed_text.parse().unwrap_or_else(|_| rand_seed());
-                                    match storage::world_meta::create_world(name, seed) {
-                                        Ok(dir) => {
-                                            // Pregen is "ready to play" when the player's working
-                                            // set of columns is loaded.
-                                            let stream_rd = graphical_core::vulkan_object::WORLD_DISTANCE;
-                                            let side = (2 * stream_rd + 1) as usize;
-                                            let total = side * side;
-                                            game_state = GameState::PreGenerating {
-                                                world_dir: dir,
-                                                seed,
-                                                loaded: 0,
-                                                total,
-                                            };
-                                        }
-                                        Err(e) => eprintln!("Failed to create world: {e}"),
-                                    }
-                                    return;
-                                }
-                                // Text input: letters go to name, digits go to seed
-                                if let Some(text) = &key_event.text {
-                                    for ch in text.chars() {
-                                        if ch.is_ascii_digit() {
-                                            if seed_text.len() < 10 {
-                                                seed_text.push(ch);
-                                            }
-                                        } else if (ch.is_ascii_alphanumeric() || ch == ' ' || ch == '-' || ch == '_') && name.len() < 24 {
-                                            name.push(ch);
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                        _ => {
-                            if key_event.state == ElementState::Pressed && key_code == KeyCode::Escape {
-                                match &game_state {
-                                    GameState::WorldSelect { .. } => game_state = GameState::TitleScreen,
-                                    GameState::TitleScreen => {
-                                        exit_program(&mut destroy_application, current_window, &mut application);
-                                    }
-                                    _ => {}
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-            Event::AboutToWait => 
-            Event::WindowEvent {
-                event: WindowEvent::RedrawRequested,
-                ..
-            } => {
-               
-            }
-            _ => (),
-        })
-        .expect("Main function crashed!");
-}
-
-fn draw_menu(ui: &mut UiPipeline, state: &mut GameState, sw: f32, sh: f32, cursor: [f32; 2], clicked: bool) {
-    match state {
-        GameState::TitleScreen => {
-            let title = "MANIFOLD";
-            let tw = UiPipeline::text_width(title, TITLE_SIZE);
-            ui.draw_text(title, (sw - tw) / 2.0, sh * 0.2, TITLE_SIZE, TEXT_COLOR);
-
-            let cx = (sw - BUTTON_W) / 2.0;
-            if draw_button(ui, "Singleplayer", cx, sh * 0.45, cursor, clicked) {
-                *state = GameState::WorldSelect {
-                    worlds: storage::world_meta::list_worlds(),
-                };
-            }
-            if draw_button(ui, "Quit", cx, sh * 0.55, cursor, clicked) {
-                std::process::exit(0);
-            }
-        }
-        GameState::WorldSelect { worlds } => {
-            let title = "Select World";
-            let tw = UiPipeline::text_width(title, TEXT_SIZE * 1.5);
-            ui.draw_text(title, (sw - tw) / 2.0, 40.0, TEXT_SIZE * 1.5, TEXT_COLOR);
-
-            let cx = (sw - BUTTON_W) / 2.0;
-            let mut y = 100.0;
-            let mut selected = None;
-            for (i, (_, meta)) in worlds.iter().enumerate() {
-                let label = format!("{} (seed: {})", meta.name, meta.seed);
-                if draw_button(ui, &label, cx, y, cursor, clicked) {
-                    selected = Some(i);
-                }
-                y += BUTTON_H + 10.0;
-            }
-            if let Some(i) = selected {
-                let (dir, meta) = &worlds[i];
-                // Existing world — enter directly, terrain streams in progressively
-                *state = GameState::EnteringWorld {
-                    world_dir: dir.clone(),
-                    seed: meta.seed,
-                };
-                return;
-            }
-
-            y += 20.0;
-            if draw_button(ui, "Create New World", cx, y, cursor, clicked) {
-                *state = GameState::CreateWorld {
-                    name: String::new(),
-                    seed_text: String::new(),
-                };
-                return;
-            }
-            if draw_button(ui, "Back", cx, y + BUTTON_H + 10.0, cursor, clicked) {
-                *state = GameState::TitleScreen;
-            }
-        }
-        GameState::CreateWorld { name, seed_text } => {
-            let title = "Create New World";
-            let tw = UiPipeline::text_width(title, TEXT_SIZE * 1.5);
-            ui.draw_text(title, (sw - tw) / 2.0, 40.0, TEXT_SIZE * 1.5, TEXT_COLOR);
-
-            let cx = (sw - BUTTON_W) / 2.0;
-            let mut y = 120.0;
-
-            ui.draw_text("Name:", cx, y, TEXT_SIZE, DIM_TEXT);
-            y += 25.0;
-            ui.draw_rect(cx, y, BUTTON_W, BUTTON_H, [0.15, 0.15, 0.2, 0.9]);
-            let name_empty = name.is_empty();
-            let display_name: &str = if name_empty { "type a name..." } else { name };
-            let name_color = if name_empty { DIM_TEXT } else { TEXT_COLOR };
-            ui.draw_text(display_name, cx + 10.0, y + 10.0, TEXT_SIZE, name_color);
-            y += BUTTON_H + 20.0;
-
-            ui.draw_text("Seed (digits, optional):", cx, y, TEXT_SIZE, DIM_TEXT);
-            y += 25.0;
-            ui.draw_rect(cx, y, BUTTON_W, BUTTON_H, [0.15, 0.15, 0.2, 0.9]);
-            let seed_empty = seed_text.is_empty();
-            let display_seed: &str = if seed_empty { "random" } else { seed_text };
-            let seed_color = if seed_empty { DIM_TEXT } else { TEXT_COLOR };
-            ui.draw_text(display_seed, cx + 10.0, y + 10.0, TEXT_SIZE, seed_color);
-            y += BUTTON_H + 30.0;
-
-            if !name.is_empty() {
-                if draw_button(ui, "Create & Play", cx, y, cursor, clicked) {
-                    let seed: u32 = seed_text.parse().unwrap_or_else(|_| rand_seed());
-                    match storage::world_meta::create_world(name, seed) {
-                        Ok(dir) => {
-                            let stream_rd = graphical_core::vulkan_object::WORLD_DISTANCE;
-                            let side = (2 * stream_rd + 1) as usize;
-                            let total = side * side;
-                            *state = GameState::PreGenerating {
-                                world_dir: dir,
-                                seed,
-                                loaded: 0,
-                                total,
-                            };
-                        }
-                        Err(e) => eprintln!("Failed to create world: {e}"),
-                    }
-                    return;
-                }
-                y += BUTTON_H + 10.0;
-            }
-            if draw_button(ui, "Back", cx, y, cursor, clicked) {
-                *state = GameState::TitleScreen;
-            }
-        }
-        GameState::PreGenerating { loaded, total, .. } => {
-            let title = if *loaded >= *total { "Finishing up..." } else { "Generating terrain..." };
-            let tw = UiPipeline::text_width(title, TEXT_SIZE * 1.5);
-            ui.draw_text(title, (sw - tw) / 2.0, sh * 0.35, TEXT_SIZE * 1.5, TEXT_COLOR);
-
-            let bar_w = 400.0;
-            let bar_h = 30.0;
-            let bx = (sw - bar_w) / 2.0;
-            let by = sh * 0.5;
-            let progress = if *total > 0 { *loaded as f32 / *total as f32 } else { 0.0 };
-            ui.draw_rect(bx, by, bar_w, bar_h, [0.15, 0.15, 0.2, 0.9]);
-            ui.draw_rect(bx + 2.0, by + 2.0, (bar_w - 4.0) * progress.min(1.0), bar_h - 4.0, [0.3, 0.7, 0.3, 1.0]);
-
-            let pct = format!("{}%", (progress * 100.0).min(100.0) as u32);
-            let pct_w = UiPipeline::text_width(&pct, TEXT_SIZE);
-            ui.draw_text(&pct, (sw - pct_w) / 2.0, by + 5.0, TEXT_SIZE, TEXT_COLOR);
-        }
-        GameState::EnteringWorld { .. } | GameState::Playing => {}
-    }
-}
-
-/// Draw a button, returns true if clicked.
-fn draw_button(ui: &mut UiPipeline, label: &str, x: f32, y: f32, cursor: [f32; 2], clicked: bool) -> bool {
-    let hovered = cursor[0] >= x && cursor[0] <= x + BUTTON_W && cursor[1] >= y && cursor[1] <= y + BUTTON_H;
-    let color = if hovered { BUTTON_HOVER } else { BUTTON_COLOR };
-    ui.draw_rect(x, y, BUTTON_W, BUTTON_H, color);
-    let tw = UiPipeline::text_width(label, TEXT_SIZE);
-    let tx = x + (BUTTON_W - tw) / 2.0;
-    let ty = y + (BUTTON_H - TEXT_SIZE) / 2.0;
-    ui.draw_text(label, tx, ty, TEXT_SIZE, TEXT_COLOR);
-    hovered && clicked
-}
-
-fn tick_pregen(state: &mut GameState, application: &mut VulkanApplication, window: &winit::window::Window, camera: &mut Camera, player: &mut Player) {
-    let GameState::PreGenerating {
-        world_dir,
-        seed,
-        loaded,
-        total,
-    } = state
-    else {
-        return;
-    };
-
-    if !application.has_world() {
-        let dir = world_dir.clone();
-        let s = *seed;
-        if let Err(e) = unsafe { application.enter_world(&dir, s) } {
-            eprintln!("Failed to enter world: {e}");
-            *state = GameState::TitleScreen;
-            return;
-        }
-    }
-    let col_count = if let Some(world) = application.world() {
-        let mut seen = std::collections::HashSet::new();
-        for cp in world.chunk_positions() {
-            seen.insert((cp.x, cp.z));
-        }
-        seen.len()
-    } else {
-        0
-    };
-    *loaded = col_count;
-    // Done when raw terrain loaded AND LOD settled (no submissions, no in-flight, stable)
-    if col_count >= *total && application.lod_settled() {
-        *state = GameState::Playing;
-        grab_cursor(window);
-        *camera = Camera::default();
-        *player = Player::new();
-    }
-}
-
-fn rand_seed() -> u32 {
-    std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .unwrap_or_default()
-        .as_millis() as u32
-}
-
-fn grab_cursor(window: &winit::window::Window) {
-    let _ = window.set_cursor_grab(CursorGrabMode::Confined);
-    window.set_cursor_visible(false);
-}
-
-fn release_cursor(window: &winit::window::Window) {
-    let _ = window.set_cursor_grab(CursorGrabMode::None);
-    window.set_cursor_visible(true);
-}
-
-fn exit_program(destroy_application: &mut bool, current_window: &EventLoopWindowTarget<()>, application: &mut VulkanApplication) {
-    *destroy_application = true;
-    current_window.exit();
-    unsafe { application.destroy_vulkan_application() }
-}
-
 fn initialize_error_handler() {
     pretty_env_logger::init();
-}
-
-struct FpsCounter {
-    frames: u32,
-    elapsed: f32,
-    display_fps: u32,
-}
-
-impl FpsCounter {
-    fn new() -> Self {
-        Self {
-            frames: 0,
-            elapsed: 0.0,
-            display_fps: 0,
-        }
-    }
-
-    fn tick(&mut self, dt: f32) {
-        self.frames += 1;
-        self.elapsed += dt;
-        if self.elapsed >= 0.5 {
-            self.display_fps = (self.frames as f32 / self.elapsed) as u32;
-            self.frames = 0;
-            self.elapsed = 0.0;
-        }
-    }
-
-    fn display(&self) -> String {
-        format!("{} fps", self.display_fps)
-    }
 }
