+++ src/utils.rs
@@ -1,0 +1,6 @@
+pub fn rand_seed() -> u32 {
+    std::time::SystemTime::now()
+        .duration_since(std::time::UNIX_EPOCH)
+        .unwrap_or_default()
+        .as_millis() as u32
+}
